refactor: 拆分设置页模型提供商管理
